整理AOP,ID相关的代码
整理AOP,ID相关的代码
