+++ src/Diary/2019/12.06.txt
@@ -1,0 +1,9 @@
+1.查了想看的书，借了部分要用的书，有了个大概的底。
+2.本来来了书，被BB阻止了，退了，以后马上要用的时候再买吧。
+3.看了Spring源码。
+4.学习了Socket客户端的内容、编译、部分的系统环境知识。
+总结：
+1.今天除了下午状态不好，整天还不错。
+2.书这里自己好好把握，优先把Spring先学好，然后就是Java基础，有时间再看操作系统基础和计网。
+3.明天继续把系统和socket看完，然后继续看Spring；
+4.把学习代码整理了，学完socket以后，可以再git开个仓库，把socket编程的内容做些小工具包装一下。
