+++ src/Java/spring/DI/DI1/SimpleContainer.java
@@ -1,4 +1,4 @@
-package Java.zhujie.DI.DI1;
+package Java.spring.DI.DI1;
 
 import java.lang.reflect.Field;
