+++ src/Java/spring/DI/DI1/SimpleInject.java
@@ -1,14 +1,11 @@
-package Java.zhujie.DI.DI2;
+package Java.spring.DI.DI1;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/**
- * 单例注解
- */
-@Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE,ElementType.FIELD})
-public @interface SimpleSingleton {
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SimpleInject {
 }
